+++ axum-extra/src/response/mod.rs
@@ -9,3 +9,6 @@ pub use erased_json::ErasedJson;
 #[cfg(feature = "json-lines")]
 #[doc(no_inline)]
 pub use crate::json_lines::JsonLines;
+
+mod wasm;
+pub use wasm::Wasm;
