+++ axum/src/response/mod.rs
@@ -64,35 +64,6 @@ impl<T> From<T> for Html<T> {
     }
 }
 
-/// A WASM response.
-///
-/// Will automatically get `Content-Type: application/wasm`.
-#[derive(Clone, Copy, Debug)]
-#[must_use]
-pub struct Wasm<T>(pub T);
-
-impl<T> IntoResponse for Wasm<T>
-where
-    T: Into<Full<Bytes>>,
-{
-    fn into_response(self) -> Response {
-        (
-            [(
-                header::CONTENT_TYPE,
-                "application/wasm",
-            )],
-            self.0.into(),
-        )
-            .into_response()
-    }
-}
-
-impl<T> From<T> for Wasm<T> {
-    fn from(inner: T) -> Self {
-        Self(inner)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::extract::Extension;
